content/html/content/test/test_bug481335.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=481335
-->
<head>
  <title>Test for Bug 481335</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=481335">Mozilla Bug 481335</a>
<p id="display">
  <a id="t">A link</a>
  <iframe id="i"></iframe>
</p>
<p id="newparent" xml:base="http://www.example.com/"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">
<![CDATA[

/** Test for Bug 481335 **/
SimpleTest.waitForExplicitFinish();
var rand = Date.now() + "-" + Math.random();

is($("t").href, "",
   "Unexpected href before set");
is($("t").href, "",
   "Unexpected cached href before set");

$("t").setAttribute("href", rand);
is($("t").href,
   window.location.href.replace(/test_bug481335.xhtml/, rand),
   "Unexpected href after set");
is($("t").href,
   window.location.href.replace(/test_bug481335.xhtml/, rand),
   "Unexpected cached href after set");
var unvisitedColor = document.defaultView.getComputedStyle($("t"), "").color;
var visitedColor;

function afterThirdLoad() {
  is(document.defaultView.getComputedStyle($("t"), "").color, visitedColor,
     "Should be visited now after third load");
  SimpleTest.finish();
}

function afterSecondLoad() {
  is(document.defaultView.getComputedStyle($("t"), "").color, visitedColor,
       "Should be visited now");
  $("t").pathname = rand;
  is(document.defaultView.getComputedStyle($("t"), "").color, visitedColor,
       "Should still be visited after setting pathname to its existing value");
  $("t").pathname += "x";
  isnot(document.defaultView.getComputedStyle($("t"), "").color, visitedColor,
       "Should not be visited after changing pathname");
  $("t").pathname = $("t").pathname;
  isnot(document.defaultView.getComputedStyle($("t"), "").color, visitedColor,
       "Should not be visited after setting unvisited pathname to existing value");

  $("i").onload = afterThirdLoad;
  $("i").src = $("t").href;
}

function afterFirstLoad() {
  visitedColor = document.defaultView.getComputedStyle($("t"), "").color;
  isnot(visitedColor, unvisitedColor, "Why are the colors the same?");
  $("newparent").appendChild($("t"));
  is($("t").href, "http://www.example.com/" + rand,
     "Unexpected href after move");
  is($("t").href, "http://www.example.com/" + rand,
     "Unexpected cached href after move");
  is(document.defaultView.getComputedStyle($("t"), "").color, unvisitedColor,
     "Should be unvisited now");

  $("i").onload = afterSecondLoad;
  $("i").src = $("t").href;
}

addLoadEvent(function() {
  $("i").onload = afterFirstLoad;
  $("i").src = $("t").href;
});
]]>
</script>
</pre>
</body>
</html>
